pyproject.toml
[tool.poetry]
name = "vancouverbeauytyguide"
version = "0.1.0"
description = ""
authors = ["Valeriy Vasilenko <hitehnik132@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.10"
python-dotenv = "^1.0.0"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"
flake8 = "^6.1.0"
pre-commit = "^3.4.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
